Move NoDiagnostic tests
